Major changes.

+++ trunk/cofmServer/src/main/java/collab/fm/server/bean/Request.java
@@ -4,19 +4,25 @@ package collab.fm.server.bean;
 public class Request extends Filterable {
 	public static final String TERMINATOR = "\n";
 	
-	private Integer id;
+	private Long id;
 	private String name;
 	private String address;
 	private String user;
+	
+	/**
+	 * Before the request is converted from JSON to Request, the JSON string is 
+	 * stored in this field.
+	 */
 	private Object data;
 	
 	public Request() {
 		
 	}
-	public Integer getId() {
+	
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getName() {
